refactor(app): move LessonClient to _components colocation (#25)
LessonClient はレッスン詳細ページ (_components) 固有の UI なので、
Next.js App Router の private folder 規約 (`_components/`) に
コロケートする。複数ページで再利用されたら `src/components/` に昇格する。

- src/components/LessonClient.tsx → _components/LessonClient.tsx
- page.tsx の import を相対パス (./_components/LessonClient) に
- hook の import も隣の _hooks への相対パスに統一
